DataGuardAI/requirements.txt
# Core ML Libraries
torch>=1.9.0
transformers>=4.21.0
datasets>=2.0.0
scikit-learn>=1.1.0
pandas>=1.4.0
numpy>=1.21.0

# NLP & Audio Processing
spacy>=3.4.0
openai-whisper>=20231117
pytesseract>=0.3.9
python-docx>=0.8.11
PyPDF2>=2.10.0
python-magic>=0.4.27
python-magic-bin>=0.4.14

# Web Framework & API
fastapi>=0.85.0
uvicorn>=0.18.0
gradio>=3.40.0
requests>=2.28.0

# Database & Storage
sqlalchemy>=1.4.0
psycopg2-binary>=2.9.0
redis>=4.3.0

# Utilities
pyyaml>=6.0
python-multipart>=0.0.5
pillow>=9.0.0
opencv-python>=4.6.0
joblib>=1.1.0
matplotlib>=3.5.0
seaborn>=0.11.0
plotly>=5.10.0

# Security & Validation
cryptography>=37.0.0
pydantic>=1.10.0

# Development
jupyter>=1.0.0
ipykernel>=6.0.0
black>=22.0.0
flake8>=4.0.0